# build/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"USAStates": {
				"enum": [
					"AL",
					"AK",
					"AS",
					"AZ",
					"AR",
					"CA",
					"CO",
					"CT",
					"DE",
					"DC",
					"FM",
					"FL",
					"GA",
					"GU",
					"HI",
					"ID",
					"IL",
					"IN",
					"IA",
					"KS",
					"KY",
					"LA",
					"ME",
					"MH",
					"MD",
					"MA",
					"MI",
					"MN",
					"MS",
					"MO",
					"MT",
					"NE",
					"NV",
					"NH",
					"NJ",
					"NM",
					"NY",
					"NC",
					"ND",
					"MP",
					"OH",
					"OK",
					"OR",
					"PW",
					"PA",
					"PR",
					"RI",
					"SC",
					"SD",
					"TN",
					"TX",
					"UT",
					"VT",
					"VI",
					"VA",
					"WA",
					"WV",
					"WI",
					"WY"
				],
				"type": "string"
			},
			"IRegistrationDetails": {
				"properties": {
					"licensePlate": {
						"type": "string"
					},
					"registrationNumber": {
						"type": "number",
						"format": "double"
					},
					"registrationState": {
						"$ref": "#/components/schemas/USAStates"
					},
					"nameOnRegistration": {
						"type": "string"
					}
				},
				"required": [
					"licensePlate",
					"registrationNumber",
					"registrationState",
					"nameOnRegistration"
				],
				"type": "object",
				"additionalProperties": false
			},
			"stdColors": {
				"enum": [
					"red",
					"yellow",
					"green",
					"blue",
					"orange",
					"white",
					"black",
					"grey",
					"silver",
					"other"
				],
				"type": "string"
			},
			"IVehicle": {
				"properties": {
					"id": {
						"type": "string"
					},
					"registration": {
						"$ref": "#/components/schemas/IRegistrationDetails"
					},
					"vinNumber": {
						"type": "number",
						"format": "double"
					},
					"ownerReportedCarValue": {
						"type": "number",
						"format": "double"
					},
					"ownerReportedCurrentMileage": {
						"type": "number",
						"format": "double"
					},
					"vehicleColor": {
						"$ref": "#/components/schemas/stdColors"
					},
					"otherColor": {
						"type": "string"
					},
					"vehicleDescription": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"registration",
					"vinNumber",
					"ownerReportedCarValue",
					"ownerReportedCurrentMileage",
					"vehicleColor",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPagedOutput_IVehicle-or-void_": {
				"properties": {
					"data": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/IVehicle"
								},
								{}
							]
						},
						"type": "array"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"perPage": {
						"type": "number",
						"format": "double"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderBy": {
				"enum": [
					"createdAt",
					"updatedAt"
				],
				"type": "string"
			},
			"OrderByDirection": {
				"enum": [
					"asc",
					"desc"
				],
				"type": "string"
			},
			"Pick_IVehicle.Exclude_keyofIVehicle.id-or-createdAt-or-updatedAt__": {
				"properties": {
					"registration": {
						"$ref": "#/components/schemas/IRegistrationDetails"
					},
					"vinNumber": {
						"type": "number",
						"format": "double"
					},
					"ownerReportedCarValue": {
						"type": "number",
						"format": "double"
					},
					"ownerReportedCurrentMileage": {
						"type": "number",
						"format": "double"
					},
					"vehicleColor": {
						"$ref": "#/components/schemas/stdColors"
					},
					"otherColor": {
						"type": "string"
					},
					"vehicleDescription": {
						"type": "string"
					}
				},
				"required": [
					"registration",
					"vinNumber",
					"ownerReportedCarValue",
					"ownerReportedCurrentMileage",
					"vehicleColor"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IVehicle.id-or-createdAt-or-updatedAt_": {
				"$ref": "#/components/schemas/Pick_IVehicle.Exclude_keyofIVehicle.id-or-createdAt-or-updatedAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			}
		},
		"securitySchemes": {
			"access_token": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "vehicle-registration-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Peter Morgan ",
			"email": "petergm@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/vehicles": {
			"get": {
				"operationId": "GetMany",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPagedOutput_IVehicle-or-void_"
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "registrationStates",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/components/schemas/USAStates"
							}
						}
					},
					{
						"in": "query",
						"name": "nameOnRegistration",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerReportedCarValueGreaterThan",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "ownerReportedCarValueLessThan",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "ownerReportedMileageGreaterThan",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "ownerReportedMileageLessThan",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "vehicleColor",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "createdAfter",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "createdBefore",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "updatedAfter",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "updatedBefore",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderBy"
						}
					},
					{
						"in": "query",
						"name": "orderByDirection",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderByDirection"
						}
					}
				]
			},
			"post": {
				"operationId": "PostOne",
				"responses": {
					"201": {
						"description": ""
					}
				},
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IVehicle.id-or-createdAt-or-updatedAt_"
							}
						}
					}
				}
			}
		},
		"/v1/vehicles/find-one": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IVehicle"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "licensePlate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "registrationNumber",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "registrationState",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "vinNumber",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/vehicles/{clientId}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IVehicle"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "clientId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}